+++ packages/odyssey-react-mui/src/DatePickers/DateField.tsx
@@ -28,8 +28,8 @@ import {
 import { DateTime } from "luxon";
 import { useTranslation } from "react-i18next";
 
-import { Field, RenderFieldComponentProps } from "../../Field";
-import { TextFieldProps } from "../../TextField";
+import { Field, RenderFieldComponentProps } from "../Field";
+import { TextFieldProps } from "../TextField";
 
 export type DateFieldProps = {
   onChange?: (value: string) => void;
