+++ packages/odyssey-react-mui/src/DatePickers/DatePicker.tsx
@@ -21,18 +21,18 @@ import {
 import { DateTime } from "luxon";
 import styled from "@emotion/styled";
 
-import { Button } from "../../Buttons";
+import { Button } from "../Buttons";
 import { DateFieldActionBar } from "./DateFieldActionBar";
 import { DateField, DateFieldProps } from "./DateField";
 import { DateFieldLocalizationProvider } from "./DateFieldLocalizationProvider";
 import { datePickerTheme } from "./datePickerTheme";
-import { FieldComponentProps } from "../../FieldComponentProps";
+import { FieldComponentProps } from "../FieldComponentProps";
 import {
   useOdysseyDesignTokens,
   DesignTokens,
-} from "../../OdysseyDesignTokensContext";
-import { OdysseyThemeProvider } from "../../OdysseyThemeProvider";
-import { TimeZonePicker } from "../TimeZonePicker";
+} from "../OdysseyDesignTokensContext";
+import { OdysseyThemeProvider } from "../OdysseyThemeProvider";
+import { TimeZonePicker } from "./TimeZonePicker";
 
 import {
   useOdysseyDateFields,
