+++ packages/odyssey-react-mui/src/DatePickers/datePickerTheme.tsx
@@ -11,7 +11,7 @@
  */
 
 import { CSSInterpolation } from "@mui/material/styles";
-import { createOdysseyMuiTheme } from "../../theme";
+import { createOdysseyMuiTheme } from "../theme";
 import { ThemeOptions } from "@mui/material";
 import * as odysseyTokens from "@okta/odyssey-design-tokens";
