+++ packages/odyssey-react-mui/src/DatePickers/index.ts
@@ -10,4 +10,6 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-export * from "./DateTimePicker";
+export * from "./DatePicker";
+export * from "./datePickerTheme";
+export * from "./TimeZonePicker";
