+++ packages/odyssey-react-mui/src/labs/DatePickers/dateTimePickerTheme.ts
@@ -14,7 +14,7 @@
 import { deepmerge } from "@mui/utils";
 import { ThemeOptions } from "@mui/material";
 
-import { datePickerTheme, dateStyles } from "./datePickerTheme";
+import { datePickerTheme, dateStyles } from "../../DatePickers/datePickerTheme";
 
 const theme: ThemeOptions = {
   components: {
